HandZone now takes a handIndex.
Hand indexes technically aren't stable identifiers because they're
based purely on the order in which hand zones were created. However,
TTS does not presently provide a way to programatically delete a
hand zone, it must be done via the editor. So hand indexes are at
least stable under normal usage, which is sufficient.