Added Zone.setRetainOccupyingObjects
This allows zones to behave similar to the way Instances behave
with respect to their zone reference i.e. when enabled the Zone
will treat objects that have left the bounds of the scripting
zone as still occupying the Zone *until* that object starts
occupying another zone.

By default this behavior is switched off.